Refactor LoginActivity and RegisterActivity to .views package
