+++ app/src/main/java/com/d/tickettoride/views/LoginActivity.kt
@@ -1,8 +1,9 @@
-package com.d.tickettoride
+package com.d.tickettoride.views
 
 import android.content.Intent
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import com.d.tickettoride.R
 import kotlinx.android.synthetic.main.activity_login.*
 
 class LoginActivity : AppCompatActivity() {
