+++ app/src/main/java/com/d/tickettoride/views/RegisterActivity.kt
@@ -1,7 +1,8 @@
-package com.d.tickettoride
+package com.d.tickettoride.views
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import com.d.tickettoride.R
 
 class RegisterActivity : AppCompatActivity() {
